move all domain services to the same directory, config import is a code smell
