+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "update_dns_records"
-version = "0.1.0"
+version = "0.2.0"
 edition = "2021"
 
 [dependencies]
@@ -8,12 +8,12 @@ serde_json = "1"
 bytes = "1"
 serde = { version = "1.0", features = ["derive"] }
 tokio = { version = "1", features = ["full"] }
-tokio-native-tls = { version = "0.3.1" }
+tokio-native-tls = { version = "0.3" }
 hyper-util = { version = "0.1", features = ["full"] }
 native-tls = "0.2"
 hyper = { version = "1", features = ["full"] }
 http-body-util = "0.1"
-rand = "0.8.5"
+rand = "0.8"
 base64 = { version = "0.21" }
 
 [features]
