--- src/toolkit/ip_services.rs
@@ -1,1 +1,0 @@
-pub type IpServices = Vec<(String, String)>;
