+++ src/toolkit/mod.rs
@@ -1,5 +1,3 @@
 pub mod config;
-pub mod domain_services;
-pub mod ip_services;
 pub mod requests;
 pub mod results;
